tests for inject cookie
